Did some refactoring & code cleaning
