+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/ui/contactcause/TextFieldSection.kt
@@ -2,13 +2,11 @@ package com.EENX15_22_17.digital_journal.android.ui.contactcause
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.Icon
 import androidx.compose.material.OutlinedTextField
 import androidx.compose.material.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
