+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/ui/suicideassessment/suicideSteps/SuicideStepsItem.kt
@@ -2,17 +2,12 @@ package com.EENX15_22_17.digital_journal.android.ui.suicideassessment
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.EENX15_22_17.digital_journal.android.ui.components_beta.getCheckBoxBtn
 import com.EENX15_22_17.digital_journal.android.ui.components_beta.CheckBoxData
-import com.EENX15_22_17.digital_journal.android.ui.theme.colorTextBlackLight
+import com.EENX15_22_17.digital_journal.android.ui.components_beta.DrawCheckBoxes
 
 @Composable
 fun SuicideStepsItem(items : List<CheckBoxData>,
@@ -38,30 +33,5 @@ fun SuicideStepsItem(items : List<CheckBoxData>,
         ) {
             DrawCheckBoxes(items = itemsListTwo)
         }
-    }
-}
-@Composable
-fun DrawCheckBoxes(items: List<CheckBoxData> ) {
-    items.forEachIndexed { index, item ->
-
-        var checkBoxState by remember {
-            mutableStateOf(false) }
-        Column() {
-            Row(
-                horizontalArrangement = Arrangement.Start,
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                getCheckBoxBtn(value = checkBoxState, onValueChange = { checkBoxState = it; item.checkBoxValue = it })
-                Text(
-                    text = " " + item.checkBoxName,
-                    style = TextStyle(fontSize = 16.sp),
-                    color = colorTextBlackLight,
-                )
-            }
-            println("Label ${item.checkBoxName} : ${item.checkBoxValue}")
-            Box(modifier = Modifier.size(width = 20.dp, height = 10.dp))
-        }
-        
-
     }
 }
